Handle overloaded forward function declarations with a new dedicated class to ensure equivalence between the calling and called functions instead of previously just matching by name.  Handle various edge cases such as implicit conversions and choice types.  Add pre compile (function header) functionality to CqlPreprocessorVisitor.  Refactor Cql2ElmVisitor to share functionality with CqlPreprocessorVisitor.  Add several new unit tests to validate the new functionality.
