+++ Src/java/cql-to-elm/src/main/java/org/cqframework/cql/cql2elm/preprocessor/FunctionDefinitionInfo.java
@@ -1,11 +1,15 @@
 package org.cqframework.cql.cql2elm.preprocessor;
 
-import org.antlr.v4.runtime.misc.Interval;
+import org.cqframework.cql.cql2elm.PreCompileOutput;
 import org.cqframework.cql.gen.cqlParser;
+
+import java.util.Objects;
+import java.util.StringJoiner;
 
 public class FunctionDefinitionInfo extends BaseInfo {
     private String name;
     private String context;
+    private PreCompileOutput preCompileOutput;
 
     public String getName() {
         return name;
@@ -13,6 +17,14 @@ public class FunctionDefinitionInfo extends BaseInfo {
 
     public void setName(String value) {
         name = value;
+    }
+
+    public void setPreCompileOutput(PreCompileOutput preCompileOutput) {
+        this.preCompileOutput = preCompileOutput;
+    }
+
+    public PreCompileOutput getPreCompileOutput() {
+        return this.preCompileOutput;
     }
 
     public String getContext() { return context; }
@@ -24,17 +36,30 @@ public class FunctionDefinitionInfo extends BaseInfo {
         return (cqlParser.FunctionDefinitionContext)super.getDefinition();
     }
 
-    public void setDefinition(cqlParser.FunctionDefinitionContext value) {
-        super.setDefinition(value);
-    }
-
     public FunctionDefinitionInfo withName(String value) {
         setName(value);
         return this;
     }
 
-    public FunctionDefinitionInfo withDefinition(cqlParser.FunctionDefinitionContext value) {
-        setDefinition(value);
-        return this;
+    @Override
+    public boolean equals(Object theO) {
+        if (this == theO) return true;
+        if (theO == null || getClass() != theO.getClass()) return false;
+        FunctionDefinitionInfo that = (FunctionDefinitionInfo) theO;
+        return Objects.equals(name, that.name) && Objects.equals(context, that.context) && Objects.equals(preCompileOutput, that.preCompileOutput);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, context, preCompileOutput);
+    }
+
+    @Override
+    public String toString() {
+        return new StringJoiner(", ", FunctionDefinitionInfo.class.getSimpleName() + "[", "]")
+                .add("name='" + name + "'")
+                .add("context='" + context + "'")
+                .add("preCompileOutput=" + preCompileOutput)
+                .toString();
     }
 }
