+++ Src/java/cql-to-elm/src/test/java/org/cqframework/cql/cql2elm/TranslationTests.java
@@ -1,6 +1,7 @@
 package org.cqframework.cql.cql2elm;
 
 import org.cqframework.cql.elm.tracking.TrackBack;
+import org.hamcrest.Matchers;
 import org.hl7.cql_annotations.r1.CqlToElmInfo;
 import org.hl7.elm.r1.*;
 import org.testng.annotations.Test;
@@ -143,6 +144,26 @@ public class TranslationTests {
 
         As as = (As)operand;
         assertThat(as.getAsTypeSpecifier(), is(instanceOf(ChoiceTypeSpecifier.class)));
+    }
+
+    @Test
+    public void testForwardDeclarationSameTypeDifferentNamespaceNormalTypes() throws IOException {
+        final CqlTranslator translator = TestUtils.createTranslator("TestForwardDeclarationSameTypeDifferentNamespaceNormalTypes.cql");
+        assertThat("Errors: " + translator.getErrors(), translator.getErrors().size(), Matchers.equalTo(0));
+
+        Library compileLibrary = translator.getTranslatedLibrary().getLibrary();
+        List<ExpressionDef> statements = compileLibrary.getStatements().getDef();
+        assertThat(statements.size(), Matchers.equalTo(3));
+    }
+
+    @Test
+    public void testForwardDeclarationSameTypeDifferentNamespaceGenericTypes() throws IOException {
+        final CqlTranslator translator = TestUtils.createTranslator("TestForwardDeclarationSameTypeDifferentNamespaceGenericTypes.cql");
+        assertThat("Errors: " + translator.getErrors(), translator.getErrors().size(), Matchers.equalTo(0));
+
+        Library compileLibrary = translator.getTranslatedLibrary().getLibrary();
+        List<ExpressionDef> statements = compileLibrary.getStatements().getDef();
+        assertThat(statements.size(), Matchers.equalTo(3));
     }
 
     // This test creates a bunch of translators on the common pool to suss out any race conditions.
