+++ Src/java/cql-to-elm/src/test/resources/org/cqframework/cql/cql2elm/LibraryTests/TestForwardDeclarationImplicitConversion.cql
@@ -1,0 +1,16 @@
+library TestForwardDeclaration
+
+// This is to test that forward declartions for overloaded functions works as expected.
+using FHIR version '4.0.0'
+
+include FHIRHelpers version '4.0.0' called FHIRHelpers
+
+parameter "Measurement Period" Interval<DateTime>
+
+context Patient
+
+define "Inpatient Encounter" :
+  ["Encounter"] Encounter
+    where LengthInDays(Encounter.period) <= 120
+
+define function "LengthInDays"(Value Interval<DateTime>): null
