+++ Src/java/cql-to-elm/src/test/resources/org/cqframework/cql/cql2elm/LibraryTests/TestForwardDeclarationLuke.cql
@@ -1,0 +1,4 @@
+library TestForwardDeclaration
+
+define function toString(value List<System.Code>): null
+define function toString(value Concept): null
