+++ Src/java/cql-to-elm/src/test/resources/org/cqframework/cql/cql2elm/TestForwardDeclarationSameTypeDifferentNamespaceGenericTypes.cql
@@ -1,0 +1,12 @@
+library TestForwardDeclarionSameTypeDifferentNamespace
+
+using FHIR version '4.0.1'
+
+define function toString(value Concept):
+  if value is null
+    then 'null'
+  else toString({ToQuantity(1), ToQuantity(2)})
+
+define function toString(value List<FHIR.Quantity>): ''
+define function toString(value List<System.Quantity>): ''
+
